Get_posts() is now sorted for event post types only. Fixes #224
When a query was targeting both event post types and other post types, it would return only matching posts of the event post types and not of the other post types. Because the _wpt_order meta key is only set for event post types.

Now we only sort queries that target event post types and event post types only.

This fixes issue #224